WIP: Move sheets data to App
