+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import configureStore from './configureStore';
 import './index.css';
-import App from './App';
+import { App } from 'modules/App';
 import registerServiceWorker from './registerServiceWorker';
 import { ThemeProvider } from 'styled-components';
 import theme from 'styles/theme';
