+++ src/modules/App/index.js
@@ -1,0 +1,3 @@
+import App from './containers/AppContainer';
+
+export { App };
